calculatewinner moved to gameReducer folder etc
