+++ react-app/src/index.js
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
 import store from './redux/store';
-import { calculateWinner } from './app/screens/Game/util';
 import App from './app';
 import styles from './app/styles/index.css';
